Implemented more context menu features and small addition to Editor's use of context menu
Various fixes

+++ FusionEngine/FusionImageLoader.h
@@ -135,7 +135,7 @@ namespace FusionEngine
 
 		struct AnimationOptions
 		{
-			float delay;
+			int delay;
 			bool loop;
 			bool pingpong;
 			bool backward;
